[ALPS03546620] gic: fix coverity issue
[Detail]
fix "out-of-bounds access" problem

MTK-Commit-Id: 8bbc42845036968b02c1794f434a447847661bb8

CR-Id: ALPS03546620
Feature: gic

+++ drivers/misc/mediatek/ext_gic/mtk-gic-v3-extend.c
@@ -479,7 +479,7 @@ char *mt_irq_dump_status_buf(int irq, char *buf)
 int mt_irq_dump_cpu(int irq)
 {
 	int rc;
-	unsigned int result;
+	unsigned long result;
 
 	irq = virq_to_hwirq(irq);
